+++ homework1/src/main/java/com/github/patbattb/hw1/commands/CompareReportsCommand.java
@@ -1,11 +1,13 @@
 package com.github.patbattb.hw1.commands;
 
+import com.github.patbattb.hw1.service.ReportHandler;
+
 /**
  * Compares data between monthly and yearly reports and prints result.
  */
 public final class CompareReportsCommand implements MenuCommand {
     @Override
-    public void run() {
-        System.out.println("CompareReportsCommand");
+    public void runCommand(ReportHandler reportHandler) {
+
     }
 }