+++ homework1/src/main/java/com/github/patbattb/hw1/commands/InfoYearlyCopmmand.java
@@ -1,11 +1,13 @@
 package com.github.patbattb.hw1.commands;
 
+import com.github.patbattb.hw1.service.ReportHandler;
+
 /**
  * Prints yearly information.
  */
 public final class InfoYearlyCopmmand implements MenuCommand {
     @Override
-    public void run() {
-        System.out.println("InfoYearlyCopmmand");
+    public void runCommand(ReportHandler reportHandler) {
+
     }
 }