+++ homework1/src/main/java/com/github/patbattb/hw1/commands/MenuCommand.java
@@ -1,7 +1,10 @@
 package com.github.patbattb.hw1.commands;
 
+import com.github.patbattb.hw1.service.ReportHandler;
+
 /**
  * Command interface.
  */
-public interface MenuCommand extends Runnable {
+public interface MenuCommand {
+    void runCommand(ReportHandler reportHandler);
 }