+++ homework1/src/main/java/com/github/patbattb/hw1/commands/ReadYearlyReportsCommand.java
@@ -1,11 +1,13 @@
 package com.github.patbattb.hw1.commands;
 
+import com.github.patbattb.hw1.service.ReportHandler;
+
 /**
  * Reads data from yearly reports.
  */
 public final class ReadYearlyReportsCommand implements MenuCommand {
     @Override
-    public void run() {
-        System.out.println("ReadYearlyReportsCommand");
+    public void runCommand(ReportHandler reportHandler) {
+
     }
 }