+++ homework1/src/main/java/com/github/patbattb/hw1/domain/Report.java
@@ -1,0 +1,7 @@
+package com.github.patbattb.hw1.domain;
+
+/**
+ * Entity contains data from downloaded reports.
+ */
+public abstract class Report {
+}