+++ homework1/src/main/java/com/github/patbattb/hw1/domain/ReportYearly.java
@@ -1,0 +1,11 @@
+package com.github.patbattb.hw1.domain;
+
+import java.util.ArrayList;
+
+/**
+ * Implementation of {@link Report} for yearly reports.
+ */
+public class ReportYearly extends Report {
+    private int date;
+    private ArrayList<TransactionYearly> transactions;
+}