+++ homework1/RELEASE_NOTES.md
@@ -1,9 +1,21 @@
 ___
-# V0.0.3
+
+# V0.2.0
+
+- Moved report files to ./reports
+- Added functionality for working with report entities
+- Implemented Read Monthly Reports Command
+
+___
+
+# V0.1.0
+
 - Added Main Menu
 - Added classes for each commands
 - Added a list of Commands
 
 ___
+
 # V0.0.2
+
 - Added checkstyle plugin